Refactor main.py into different files and add compatibility for more providers like OpenAI
Refactor `main.py` into multiple files and add compatibility for OpenAI.

* **Refactoring:**
  - Add `model_loader.py` with functions `load_model` and `does_chat_fit_in_context`.
  - Add `web_crawler.py` with functions `crawl` and `crawl_sync`.
  - Add `wikipedia_tools.py` with functions `wikipedia_search` and `wikipedia`.
  - Add `chat_interface.py` with the main chat interface logic and tool definitions.
  - Modify `main.py` to import functions from the new files and update the `main` function.

* **OpenAI Compatibility:**
  - Add `openai_provider.py` with functions `load_openai_model` and `interact_with_openai_model`.
  - Update `README.md` to include instructions for using OpenAI as a provider.
  - Update `requirements.txt` to include the OpenAI dependency.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Dariton4000/agent?shareId=XXXX-XXXX-XXXX-XXXX).
